fix(document-request): cast newStatus a int per check strict in_array
canUpdateStatus usa in_array(..., true) strict, ma il POST 'status'
arriva come stringa: '1' !== 1 (int) e ogni cambio stato veniva
rifiutato con 403 anche per super_admin.

Cast esplicito a int subito dopo la lettura dal POST.

+++ frontend/controllers/DocumentRequestController.php
@@ -115,7 +115,7 @@ class DocumentRequestController extends BaseController
         }
         
         $model = $this->findModel($id);
-        $newStatus = Yii::$app->request->post('status');
+        $newStatus = (int) Yii::$app->request->post('status');
         
         // Log per debugging
         Yii::info("Update status request - ID: $id, Current Status: {$model->status}, New Status: $newStatus", __METHOD__);
